revamp db functions, subscriptionExpenses

+++ src/services/firebase/admin-delete.ts
@@ -1,17 +1,26 @@
 "use server";
 
-import { getAuth } from "firebase-admin/auth";
-import { firestore } from "firebase-admin";
+// Local Imports
 import { firestoreAdmin } from "@/lib/firebase/config-admin";
-import { retrieveUserIdAdmin } from "./retrieve-admin";
+import { retrieveUIDAdmin } from "./admin-retrieve";
+
+// External Imports
+import { firestore } from "firebase-admin";
+import { getAuth } from "firebase-admin/auth";
+import { expensesCol, inventoryCol, ordersCol, usersCol } from "./constants";
 
 
 async function deleteUserAdmin({ idToken }: { idToken: string }): Promise<{ success?: boolean; error?: string }> {
     try {
-        const uid = await retrieveUserIdAdmin({ idToken });
-
+        // Step 1: Retrieve UID
+        const uid = await retrieveUIDAdmin({ idToken });
+        if (!uid) throw Error("User could not be found");
+        
+        // Step 2: Delete users documents
         const { error: deleteDocError } = await deleteUserDocsAdmin({ uid });
         if (deleteDocError) throw deleteDocError;
+        
+        // Step 3: Delete users authentication
         const { error: deleteAuthError } = await deleteAuthUserAdmin({ uid });
         if (deleteAuthError) throw deleteAuthError;
 
@@ -33,15 +42,15 @@ async function deleteAuthUserAdmin({ uid }: { uid: string }): Promise<{ success?
 
 async function deleteUserDocsAdmin({ uid }: { uid: string }): Promise<{ success?: boolean; error?: any }> {
     try {
-        const userDocRef = firestoreAdmin.collection("users").doc(uid);
-        const inventoryDocRef = firestoreAdmin.collection("inventory").doc(uid);
-        const orderDocRef = firestoreAdmin.collection("orders").doc(uid);
-        const expensesDocRef = firestoreAdmin.collection("expenses").doc(uid);
-        
+        const userDocRef = firestoreAdmin.collection(usersCol).doc(uid);
+        const orderDocRef = firestoreAdmin.collection(ordersCol).doc(uid);
+        const expensesDocRef = firestoreAdmin.collection(expensesCol).doc(uid);
+        const inventoryDocRef = firestoreAdmin.collection(inventoryCol).doc(uid);
+
         await userDocRef.delete();
-        await firestore().recursiveDelete(inventoryDocRef);
         await firestore().recursiveDelete(orderDocRef);
         await firestore().recursiveDelete(expensesDocRef);
+        await firestore().recursiveDelete(inventoryDocRef);
 
         return { success: true };
     } catch (error) {
